Seed up BDF import in Poseidon.
git-svn-id: https://svnsrv.germanlloyd.org/svn/ExtFEMIO/trunk@261 389f8db8-dd28-11e4-b689-8b400cb78fb4

+++ src/fem/fem_cards_bnbcd.cpp
@@ -67,7 +67,7 @@ namespace dnvgl {
             bnbcd::bnbcd(
                const long &NODENO,
                const ::std::deque<long> &FIX) :
-               card(), NODENO(NODENO), NDOF(FIX.size()), FIX(FIX) {}
+               card(), NODENO(NODENO), NDOF((long)FIX.size()), FIX(FIX) {}
 
             const ::dnvgl::extfem::fem::cards::types
             bnbcd::card_type(void) const {return BNBCD;}
